+++ Pesto/include/EditorLayer.h
@@ -23,4 +23,6 @@ class EditorLayer : public Basil::Layer
 		Basil::Shared<Basil::Framebuffer> framebuffer;
 		glm::vec4 squareColor = { 0.2f, 0.3f, 0.8f, 1.0f };
 		glm::vec2 viewportSize = { 0.0f, 0.0f };
+		bool viewportFocused;
+		bool viewportHovered;
 };